Converted ScrollToTop component to TS and fixed a couple bugs
