+++ src/app/about/about.css
@@ -80,7 +80,6 @@ section {
   z-index: 1;
 }
 
-
 .about-images .img1 {
   width: 400px;
   border-radius: 8px;
@@ -192,8 +191,6 @@ section {
   transform: scale(0.95);
 }
 
-
-
 /* Partner section   */
 .partner-section {
   background-color: var(--slateBlue);
@@ -223,7 +220,7 @@ section {
   top: 0;
   width: 150px;
   height: 100%;
-  content: "";
+  content: '';
   z-index: 2;
 }
 
@@ -292,16 +289,13 @@ section {
   gap: 50px;
   align-items: center;
   justify-content: space-between;
-  width: 100vw;
-  max-width: 100vw;
+  width: 100%;
   margin: 0;
   padding-left: 50px;
   padding-right: 50px;
   padding-top: 80px;
   padding-bottom: 80px;
   box-sizing: border-box;
-  margin-left: "calc(50% - 50vw)";
-  margin-right: "calc(50% - 50vw)",
 }
 
 .why-left {
@@ -346,7 +340,6 @@ section {
 .years-badge span {
   font-size: 14px;
 }
-
 
 .why-left h6 {
   color: #00447c;
@@ -395,10 +388,7 @@ section {
 
 .icon {
   margin-right: 10px;
-
 }
-
-
 
 /* CTA Section */
 .cta-section {
@@ -417,8 +407,11 @@ section {
   top: 0;
   left: 0;
   position: absolute;
-  background: linear-gradient(0deg, rgba(1, 31, 68, 0.60), rgba(1, 31, 68, 0.60));
-  transition: background 0.3s, border-radius 0.3s, opacity 0.3s;
+  background: linear-gradient(0deg, rgba(1, 31, 68, 0.6), rgba(1, 31, 68, 0.6));
+  transition:
+    background 0.3s,
+    border-radius 0.3s,
+    opacity 0.3s;
 }
 
 .r-container {
@@ -518,7 +511,6 @@ p {
     padding: 12px 24px;
   }
 }
-
 
 /* Team */
 .team-cards-section {
@@ -635,7 +627,6 @@ p {
   margin-bottom: 40px;
 }
 
-
 .all-service-btn {
   position: relative;
   display: inline-block;
@@ -709,7 +700,6 @@ p {
   transition: transform 0.3s ease;
 }
 
-
 .service-card i {
   font-size: 36px;
   color: var(--slateBlue);
@@ -755,7 +745,7 @@ p {
   display: inline-flex;
   align-items: center;
   gap: 8px;
-  background-color: #011F44;
+  background-color: #011f44;
   color: #fff;
   padding: 12px 24px;
   border-radius: 5px;
@@ -764,8 +754,8 @@ p {
 }
 
 .learn-more:hover {
-  background-color: #FEC28E;
-  color: #011F44;
+  background-color: #fec28e;
+  color: #011f44;
   transform: translateY(-3px);
 }
 
@@ -780,7 +770,7 @@ p {
 }
 
 .learn-more:hover {
-  background-color: #FFD28F;
+  background-color: #ffd28f;
   color: var(--slateBlue);
 }
 
@@ -825,9 +815,7 @@ p {
   }
 }
 
-
 @media (max-width: 900px) {
-
   .about-container,
   .why-section {
     flex-direction: column;
@@ -880,7 +868,7 @@ p {
   background: #fff;
   padding: 30px 20px;
   width: 300px;
-  height:480px;
+  height: 480px;
   border-radius: 10px;
   box-shadow: 0 4px 20px rgba(0, 0, 0, 0.08);
   transition: transform 0.3s ease;
@@ -948,7 +936,7 @@ p {
   background: var(--gold);
   color: #222;
 }
-.founder-podcast-section{
+.founder-podcast-section {
   text-align: center;
 }
