+++ src/components/ScrollToTop.tsx
@@ -1,0 +1,44 @@
+'use client';
+import { useState, useEffect } from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faArrowUp } from '@fortawesome/free-solid-svg-icons';
+
+const ScrollToTopButton = () => {
+  const [isVisible, setIsVisible] = useState(false);
+  const [lastScrollY, setLastScrollY] = useState(0); // Track last scroll position
+
+  useEffect(() => {
+    const handleScroll = () => {
+      const currentScrollY = window.scrollY;
+
+      if (
+        (currentScrollY > 300 && currentScrollY > lastScrollY) ||
+        currentScrollY + window.innerHeight >= document.body.offsetHeight - 1
+      ) {
+        setIsVisible(true);
+      } else setIsVisible(false);
+
+      setLastScrollY(currentScrollY); // Update last scroll position
+    };
+
+    window.addEventListener('scroll', handleScroll);
+    return () => window.removeEventListener('scroll', handleScroll);
+  }, [lastScrollY]); // Depend on lastScrollY to detect changes
+
+  const scrollToTop = () => {
+    window.scrollTo({ top: 0, behavior: 'smooth' });
+  };
+
+  return (
+    <button
+      onClick={scrollToTop}
+      className={`fixed bottom-6 z-[33] right-6 bg-blue-500 text-white p-4 rounded-full shadow-lg flex items-center justify-center text-2xl transition-opacity duration-300 hover:bg-blue-600 ${
+        isVisible ? 'opacity-100' : 'opacity-0 pointer-events-none'
+      }`}
+    >
+      <FontAwesomeIcon icon={faArrowUp} />
+    </button>
+  );
+};
+
+export default ScrollToTopButton;
